backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Marty Sullivan <marty@martysullivan.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
aioboto3 = "<13"
langchain_experimental = "<1"
langserve = {extras = ["server"], version = ">=0.0.30"}
openai = "<2"
pydantic = "<2"
uvicorn = "^0.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
